refacto: move GradeComparator to its own file
